--- src/sentry/notifications/platform/providers/__init__.py
@@ -1,11 +1,0 @@
-__all__ = [
-    "DiscordNotificationProvider",
-    "EmailNotificationProvider",
-    "MSTeamsNotificationProvider",
-    "SlackNotificationProvider",
-]
-
-from .discord import DiscordNotificationProvider
-from .email import EmailNotificationProvider
-from .msteams import MSTeamsNotificationProvider
-from .slack import SlackNotificationProvider
